+++ frontend/src/specs/hello.spec.js
@@ -1,8 +1,0 @@
-import getMessage from '../hello';
-
-test('has default message', () => {
-  const message = getMessage();
-  expect(message).toEqual(
-    'Welcome to your React site, built with minimal dependencies.'
-  );
-});