+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add React static Home and About pages, with navigation and routing
 - Add theme and styling
+- Add React Testing Library and Web Test Runner. Remove Jest.
 
 ## [0.0.1] - 2021-10-03
 ### Added